requirements.txt
python-dotenv
openai
pydantic
swarm
pytest
instructor
qdrant-client
pandas
llama-index-llms-openai
llama-index